1. 403 에러가 간헐적으로 발생. 이미지를 다운로드 하지 않고 다음'화'로 넘어감. -> 엑박이 있는 경우가 대개 그랬음.    connection 다음에 addRequestProperty("User-Agent", "Mozilla/4.76"); 를 추가.
2. 다운로드 중 속도가 현저히 느려짐 + 범위오류 발생
-> byte[] buf = new byte[conn.getContentLength()]; 를 널널하게
   byte[] buf = new byte[1024];로 수정.

3. 완료 메시지 생성시 스크롤 자동 이동
4. 이미 있는 파일의 경우 skip (받다가 중지했을 경우, 다음에 똑같이 실행시키면 다음 파일부터 . 5. jpg파일만 받을
수 있던 걸 jpeg,png,gif파일까지 받게 수정.

+++ maurmaur_P/src/maurmaur_P/downloader.java
@@ -19,12 +19,14 @@ import org.jsoup.nodes.Element;
 import org.jsoup.select.Elements;
 
 public class downloader implements Runnable {
+	private Object textLog;
+	String ThreadName;
 	List<String> comicList = new ArrayList<String>();
 	List<String> path = new ArrayList<String>();
 	List<String> namae = new ArrayList<String>();
-	private Object textLog;
 	JTextArea txtLog = new JTextArea();
-	String ThreadName;
+	String format, confirm;
+	File f;
 
 	downloader(List<String> comicList, List<String> path, List<String> namae, JTextArea txtLog, String ThreadName) {
 		this.comicList = comicList;
@@ -36,47 +38,63 @@ public class downloader implements Runnable {
 
 	@Override
 	public void run() {
+		long start = System.currentTimeMillis();
 		for (int i = 0; i < comicList.size(); i++) {
 			try {// 한 권의 img를 저장하는 try
 				Document doc2 = Jsoup.connect(comicList.get(i)).get();
 				Element element2 = doc2.select("div.article-gallery").get(0);
 				Elements img = element2.select("img");
-				// System.out.println(comicList.get(i));
-
 				int fileNum = 1;
 				for (Element e2 : img) {
 
-					String url3 = e2.getElementsByAttribute("src").attr("data-src");
-
-					url3 = "http://wasabisyrup.com" + url3;
-
+					String url3 = "http://wasabisyrup.com" + e2.getElementsByAttribute("src").attr("data-src");
 					url3 = url3.replace(" copy", "%20copy");
-
 					URL imgUrl = new URL(url3);
 
-					String format = ".jpg";
-
-					HttpURLConnection conn = (HttpURLConnection) imgUrl.openConnection();
-					System.out.println(conn.getContentLength());
-
-					InputStream is = conn.getInputStream();
-					BufferedInputStream bis = new BufferedInputStream(is);
-					FileOutputStream os = new FileOutputStream(
-							path.get(i) + "\\" + namae.get(i) + "_" + fileNum + format);
-
-					BufferedOutputStream bos = new BufferedOutputStream(os);
-					int byteImg;
-
-					byte[] buf = new byte[conn.getContentLength()];
-					while ((byteImg = bis.read(buf)) != -1) {
-						bos.write(buf, 0, byteImg);
+					if (url3.contains(".jpg")) {
+						format = ".jpg";
+					} else if (url3.contains(".jpeg")) {
+						format = ".jpeg";
+					} else if (url3.contains(".gif")) {
+						format = ".gif";
+					} else if (url3.contains(".png")) {
+						format = ".png";
 					}
-					txtLog.append(namae.get(i) + "_" + fileNum + " 저장완료" + "  src : " + imgUrl + "\n");
-					fileNum += 1;
-					bos.close();
-					os.close();
-					bis.close();
-					is.close();
+
+					// 이미 있는 파일이라면 넘어감.
+					confirm = path.get(i) + "\\" + namae.get(i) + "_" + fileNum + format;
+					f = new File(confirm);
+					if (f.exists()) {
+						System.out.println(confirm + "은 이미 존재해서 넘어감");
+					}
+					// 이미지 다운로드
+					else {
+						HttpURLConnection conn = (HttpURLConnection) imgUrl.openConnection();
+						conn.addRequestProperty("User-Agent", "Mozilla/4.76");
+						System.out.println(conn.getContentLength());
+
+						InputStream is = conn.getInputStream();
+						BufferedInputStream bis = new BufferedInputStream(is);
+						FileOutputStream os = new FileOutputStream(
+								path.get(i) + "\\" + namae.get(i) + "_" + fileNum + format);
+
+						BufferedOutputStream bos = new BufferedOutputStream(os);
+						int byteImg;
+
+						// byte[] buf = new byte[conn.getContentLength()]; // 범위오류 때문에 1024byte단위로 읽게
+						// 수정.
+						byte[] buf = new byte[1024];
+						while ((byteImg = bis.read(buf)) != -1) {
+							bos.write(buf, 0, byteImg);
+						}
+						txtLog.append(namae.get(i) + "_" + fileNum + " 저장완료" + "  src : " + imgUrl + "\n");
+						txtLog.setCaretPosition(txtLog.getDocument().getLength());
+						fileNum += 1;
+						bos.close();
+						os.close();
+						bis.close();
+						is.close();
+					}
 
 				}
 
@@ -86,7 +104,9 @@ public class downloader implements Runnable {
 			}
 		}
 
-		txtLog.append(ThreadName + " is Done\n");
+		long end = System.currentTimeMillis();
+		txtLog.append(ThreadName + " is Done ____ " + "실행 시간 : " + (end - start) / 1000.0 + "초\n");
+		txtLog.setCaretPosition(txtLog.getDocument().getLength());
 
 	}
